disabled_denorm_refactor
Previously the team set pruner was designed to work only with an instance that had enabled denormalization. However, this isn't even the default beahvior for sugar, so assuming it's enabled isn't safe.

I chagned all of the SQL queries that depended on the denorm table use the team_sets table instead. The data that we need is actually team set id's, so using this table instead of the active denorm table not only makes more sense, but should also be more performant. It wouldn't have worked for JPMC because they managed to delete a huge, more-or-less random chunk from their team_sets table.

I also changed the backup and restore methods to only work with the denorm table if denormalization is enabled.

I removed the preflight check that would stop the prune if denormalization wasn't enabled. It still checks for the active table being available and the system not being in a rebuild state.

I changed the help messages in the commands to reflect that the pruner was no longer working off of the denorm table.